typing and code org
