+++ front/lib/providers.ts
@@ -167,56 +167,3 @@ export async function getProviderLLMModels(
   let models = await modelsRes.json();
   return { models: models.models };
 }
-
-type Credentials = {
-  OPENAI_API_KEY?: string;
-  COHERE_API_KEY?: string;
-  AI21_API_KEY?: string;
-  AZURE_OPENAI_API_KEY?: string;
-  AZURE_OPENAI_ENDPOINT?: string;
-  ANTHROPIC_API_KEY?: string;
-  SERP_API_KEY?: string;
-  SERPER_API_KEY?: string;
-  BROWSERLESS_API_KEY?: string;
-};
-
-export const credentialsFromProviders = (
-  providers: { providerId: string; config: string }[]
-): Credentials => {
-  let credentials: Credentials = {};
-
-  providers.forEach((provider) => {
-    let config = JSON.parse(provider.config) as {
-      api_key: string;
-      endpoint?: string;
-    };
-
-    switch (provider.providerId) {
-      case "openai":
-        credentials["OPENAI_API_KEY"] = config.api_key;
-        break;
-      case "cohere":
-        credentials["COHERE_API_KEY"] = config.api_key;
-        break;
-      case "ai21":
-        credentials["AI21_API_KEY"] = config.api_key;
-        break;
-      case "azure_openai":
-        credentials["AZURE_OPENAI_API_KEY"] = config.api_key;
-        credentials["AZURE_OPENAI_ENDPOINT"] = config.endpoint;
-        break;
-      case "anthropic":
-        credentials["ANTHROPIC_API_KEY"] = config.api_key;
-      case "serpapi":
-        credentials["SERP_API_KEY"] = config.api_key;
-        break;
-      case "serper":
-        credentials["SERPER_API_KEY"] = config.api_key;
-        break;
-      case "browserlessapi":
-        credentials["BROWSERLESS_API_KEY"] = config.api_key;
-        break;
-    }
-  });
-  return credentials;
-};
