+++ front/package.json
@@ -15,7 +15,6 @@
     "@heroicons/react": "^2.0.11",
     "@sendgrid/mail": "^7.7.0",
     "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/line-clamp": "^0.4.4",
     "@uiw/react-textarea-code-editor": "^2.0.6",
     "ajv": "^8.12.0",
     "blake3": "^2.1.7",
