+++ front/types/provider.ts
@@ -2,3 +2,15 @@ export type ProviderType = {
   providerId: string;
   config: string;
 };
+
+export type CredentialsType = {
+  OPENAI_API_KEY?: string;
+  COHERE_API_KEY?: string;
+  AI21_API_KEY?: string;
+  AZURE_OPENAI_API_KEY?: string;
+  AZURE_OPENAI_ENDPOINT?: string;
+  ANTHROPIC_API_KEY?: string;
+  SERP_API_KEY?: string;
+  SERPER_API_KEY?: string;
+  BROWSERLESS_API_KEY?: string;
+};
